+++ extensions/brew/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Brew Changelog
 
+## [Improvements] - {PR_MERGE_DATE}
+
+- Optimized checkmark icon
+- Updated metadata
+
 ## [Improvements] - 2025-01-02
 
 - Add terminal option for Ghostty